fixed root user issue

+++ install/install.sh
@@ -1,31 +1,42 @@
 #!/bin/bash
 
+# Don't run the script as privileged user
+if [[ $EUID -eq 0 ]]; then
+   echo "You must not run the script as root or privileged user." 
+   exit 1
+fi
+
+
 # Install aria2
 sudo apt update
 sudo apt install aria2 -y
+
+
+
 
 # source directory of bash script
 DIR="$( cd "$( dirname "${BASH_SOURCE[0]}" )" >/dev/null 2>&1 && pwd )"
 
 # aria2 configuration
-mkdir $HOME/.aria2
-cp $DIR/aria2.conf $HOME/.aria2/aria2.conf
+[ ! -d $HOME/.aria2 ] && mkdir $HOME/.aria2
+[ ! -f $HOME/.aria2/aria2.conf ] && cp $DIR/aria2.conf $HOME/.aria2/aria2.conf
 
 # aria2c startup daemon
-mkdir $HOME/.init
-cp $DIR/aria2_init.conf $HOME/.init/aria2.conf
+[ ! -d $HOME/.init ] && mkdir $HOME/.init
+[ ! -f $HOME/.init/aria2.conf ] && cp $DIR/aria2_init.conf $HOME/.init/aria2.conf
 
 # aria2c appliation menu desktop entry
 DSKFILE="$HOME/.local/share/applications/aria2-webui.desktop"
+rm $DSKFILE
 cat > $DSKFILE << ENDOFFILE
 [Desktop Entry]
 Encoding=UTF-8
 Name=Aria2 WebUI
-Type=Link
+Type=Application
 Terminal=false
 Categories=Network
 ENDOFFILE
-echo "URL=file://$DIR/../docs/index.html" >> $DSKFILE
+echo "Exec=xdg-open $DIR/../docs/index.html" >> $DSKFILE
 echo "Icon=$DIR/Network-Download-icon.png" >> $DSKFILE
 
 sudo chmod +x $DSKFILE
